:green_heart: Reduce size of build file
**Problem**
The grunt file is quiet large (~600 lines) which make it difficult to
understand an debug.

**Solution**
* Abstract every task configuration in `grunt/`
* Create `grunt/aliases.js` to store tasks

**Result**
Gruntfile contains less than 100 lines
