+++ Gruntfile.js
@@ -1,555 +1,54 @@
-// Generated on 2015-07-26 using generator-angular 0.12.1
 'use strict'
 
 module.exports = function(grunt) {
-
   // Time how long tasks take. Can help when optimizing build times
   require('time-grunt')(grunt)
 
   // Automatically load required Grunt tasks
-  require('jit-grunt')(grunt, {
-    useminPrepare: 'grunt-usemin',
-    ngtemplates: 'grunt-angular-templates',
-    cdnify: 'grunt-google-cdn',
-    coveralls: 'grunt-karma-coveralls',
-    jsdoc: 'grunt-jsdoc',
-    protractor: 'grunt-protractor-runner'
-  })
-
-  // Configurable paths for the application
-  var appConfig = {
-    app: require('./bower.json').appPath || 'app',
-    dist: 'dist'
-  }
-
-  // Define the configuration for all the tasks
-  grunt.initConfig({
-
-    // Project settings
-    yeoman: appConfig,
-
-    // Watches files for changes and runs tasks based on the changed files
-    watch: {
-      bower: {
-        files: ['bower.json'],
-        tasks: ['wiredep']
+  var config = require('load-grunt-config')(grunt, {
+		data: {
+      path: {
+        app: 'app',
+        dist: 'dist',
+        test: 'test',
+        doc: 'dist/doc',
+        karmaConf: 'test/karma.conf.js'
       },
-      js: {
-        files: [
-          '<%= yeoman.app %>/scripts/**/*.js',
-          '<%= yeoman.app %>/components/**/*.js'
-        ],
-        tasks: ['newer:jshint:all', 'newer:jscs'],
-        options: {
-          livereload: '<%= connect.options.livereload %>'
-        }
-      },
-      jsTest: {
-        files: ['test/unit/**/*.js'],
-        tasks: ['newer:jshint:test', 'newer:jscs', 'karma']
-      },
-      compass: {
-        files: ['<%= yeoman.app %>/styles/**/*.{scss,sass}'],
-        tasks: ['compass:server', 'autoprefixer:server']
-      },
-      gruntfile: {
-        files: ['Gruntfile.js']
-      },
-      livereload: {
-        options: {
-          livereload: '<%= connect.options.livereload %>'
-        },
-        files: [
-          '<%= yeoman.app %>/**/*.html',
-          '.tmp/styles/**/*.css',
-          '<%= yeoman.app %>/images/**/*.{png,jpg,jpeg,gif,webp,svg}'
-        ]
-      }
-    },
-
-    jsdoc: {
-      dist: {
-        src: [
-          'app/scripts/**/*.js',
-          'app/components/**/*.js'
-        ],
-        options: {
-          destination: appConfig.dist + '/doc',
-          readme: 'README.md',
-          template: 'node_modules/ink-docstrap/template',
-          configure: 'jsdoc.conf.json'
-        }
-      }
-    },
-
-    // The actual grunt server settings
-    connect: {
       options: {
         port: 9000,
-        // Change this to '0.0.0.0' to access the server from outside.
         hostname: 'localhost',
-        livereload: 35729
-      },
-      livereload: {
-        options: {
-          open: true,
-          middleware: function(connect) {
-            return [
-              connect.static('.tmp'),
-              connect().use(
-                '/bower_components',
-                connect.static('./bower_components')
-              ),
-              connect().use(
-                '/app/styles',
-                connect.static('./app/styles')
-              ),
-              connect.static(appConfig.app)
-            ]
-          }
-        }
-      },
-      test: {
-        options: {
-          port: 9001,
-          middleware: function(connect) {
-            return [
-              connect.static('.tmp'),
-              connect.static('test'),
-              connect().use(
-                '/bower_components',
-                connect.static('./bower_components')
-              ),
-              connect.static(appConfig.app)
-            ]
-          }
-        }
-      },
-      dist: {
-        options: {
-          open: true,
-          base: '<%= yeoman.dist %>'
-        }
-      }
-    },
-
-    // Make sure code styles are up to par and there are no obvious mistakes
-    jshint: {
-      options: {
-        jshintrc: '.jshintrc',
-        reporter: require('jshint-stylish')
-      },
-      all: {
-        src: [
-          'Gruntfile.js',
-          '<%= yeoman.app %>/components/**/*.js',
-          '<%= yeoman.app %>/scripts/**/*.js'
-        ]
-      },
-      test: {
-        options: {
-          jshintrc: '.jshintrc'
-        },
-        src: ['test/**/*.js']
-      }
-    },
-
-    jscs: {
-      src: [
-        'Gruntfile.js',
-        '<%= yeoman.app %>/components/**/*.js',
-        '<%= yeoman.app %>/scripts/**/*.js',
-        'test/**/*.js'
-      ],
-      options: {
-        config: '.jscsrc'
-      }
-    },
-
-    versioncheck: {
-      options: {
-        hideUpToDate: false
-      }
-    },
-
-    // Empties folders to start fresh
-    clean: {
-      dist: {
-        files: [{
-          dot: true,
-          src: [
-            '.tmp',
-            '<%= yeoman.dist %>/**/*',
-            '!<%= yeoman.dist %>/.git**/*'
-          ]
-        }]
-      },
-      server: '.tmp'
-    },
-
-    // Add vendor prefixed styles
-    autoprefixer: {
-      options: {
-        browsers: ['last 1 version']
-      },
-      server: {
-        options: {
-          map: true
-        },
-        files: [{
-          expand: true,
-          cwd: '.tmp/styles/',
-          src: '**/*.css',
-          dest: '.tmp/styles/'
-        }]
-      },
-      dist: {
-        files: [{
-          expand: true,
-          cwd: '.tmp/styles/',
-          src: '**/*.css',
-          dest: '.tmp/styles/'
-        }]
-      }
-    },
-
-    // Automatically inject Bower components into the app
-    wiredep: {
-      app: {
-        src: ['<%= yeoman.app %>/index.html'],
-        ignorePath:  /\.\.\//
-      },
-      test: {
-        devDependencies: true,
-        src: '<%= karma.unit.configFile %>',
-        ignorePath:  /\.\.\//,
-        fileTypes:{
-          js: {
-            block: /(([\s\t]*)\/{2}\s*?bower:\s*?(\S*))(\n|\r|.)*?(\/{2}\s*endbower)/gi,
-              detect: {
-                js: /'(.*\.js)'/gi
-              },
-              replace: {
-                js: '\'{{filePath}}\','
-              }
-            }
-          }
-      },
-      sass: {
-        src: ['<%= yeoman.app %>/styles/**/*.{scss,sass}'],
-        ignorePath: /(\.\.\/){1,2}bower_components\//
-      }
-    },
-
-    // Compiles Sass to CSS and generates necessary files if requested
-    compass: {
-      options: {
-        sassDir: '<%= yeoman.app %>/styles',
-        cssDir: '.tmp/styles',
-        generatedImagesDir: '.tmp/images/generated',
-        imagesDir: '<%= yeoman.app %>/images',
-        javascriptsDir: '<%= yeoman.app %>/scripts',
-        fontsDir: '<%= yeoman.app %>/styles/fonts',
-        importPath: './bower_components',
-        httpImagesPath: '/images',
-        httpGeneratedImagesPath: '/images/generated',
-        httpFontsPath: '/styles/fonts',
-        relativeAssets: false,
-        assetCacheBuster: false,
-        raw: 'Sass::Script::Number.precision = 10\n'
-      },
-      dist: {
-        options: {
-          generatedImagesDir: '<%= yeoman.dist %>/images/generated'
-        }
-      },
-      server: {
-        options: {
-          sourcemap: true
-        }
-      }
-    },
-
-    // Renames files for browser caching purposes
-    filerev: {
-      dist: {
-        src: [
-          '<%= yeoman.dist %>/scripts/**/*.js',
-          '<%= yeoman.dist %>/components/**/*.js',
-          '<%= yeoman.dist %>/styles/**/*.css',
-          '<%= yeoman.dist %>/images/**/*.{png,jpg,jpeg,gif,webp,svg}',
-          '<%= yeoman.dist %>/styles/fonts/*'
-        ]
-      }
-    },
-
-    // Reads HTML for usemin blocks to enable smart builds that automatically
-    // concat, minify and revision files. Creates configurations in memory so
-    // additional tasks can operate on them
-    useminPrepare: {
-      html: '<%= yeoman.app %>/index.html',
-      options: {
-        dest: '<%= yeoman.dist %>',
-        flow: {
-          html: {
-            steps: {
-              js: ['concat'],
-              css: ['cssmin']
-            },
-            post: {}
-          }
-        }
-      }
-    },
-
-    // Performs rewrites based on filerev and the useminPrepare configuration
-    usemin: {
-      html: ['<%= yeoman.dist %>/**/*.html'],
-      css: ['<%= yeoman.dist %>/styles/**/*.css'],
-      js: [
-        '<%= yeoman.dist %>/scripts/**/*.js',
-        '<%= yeoman.dist %>/components/**/*.js'
-      ],
-      options: {
-        assetsDirs: [
-          '<%= yeoman.dist %>',
-          '<%= yeoman.dist %>/images',
-          '<%= yeoman.dist %>/styles'
-        ],
-        patterns: {
-          js: [[/(images\/[^''""]*\.(png|jpg|jpeg|gif|webp|svg))/g,
-          'Replacing references to images']]
-        }
-      }
-    },
-
-    // The following *-min tasks will produce minified files in the dist folder
-    // By default, your `index.html`'s <!-- Usemin block --> will take care of
-    // minification. These next options are pre-configured if you do not wish
-    // to use the Usemin blocks.
-    cssmin: {
-      dist: {
-        files: {
-          '<%= yeoman.dist %>/styles/main.css': [
-            '.tmp/styles/**/*.css'
-          ]
-        }
-      }
-    },
-    // uglify: {
-    //   dist: {
-    //     files: {
-    //       '<%= yeoman.dist %>/scripts/scripts.js': [
-    //         '<%= yeoman.dist %>/scripts/scripts.js'
-    //       ]
-    //     }
-    //   }
-    // },
-    concat: {
-      dist: {}
-    },
-
-    imagemin: {
-      dist: {
-        files: [{
-          expand: true,
-          cwd: '<%= yeoman.app %>/images',
-          src: '**/*.{png,jpg,jpeg,gif}',
-          dest: '<%= yeoman.dist %>/images'
-        }]
-      }
-    },
-
-    svgmin: {
-      dist: {
-        files: [{
-          expand: true,
-          cwd: '<%= yeoman.app %>/images',
-          src: '**/*.svg',
-          dest: '<%= yeoman.dist %>/images'
-        }]
-      }
-    },
-
-    htmlmin: {
-      dist: {
-        options: {
+        livereload: 35729,
+        htmlmin: {
           collapseWhitespace: true,
           conservativeCollapse: true,
           collapseBooleanAttributes: true,
           removeCommentsFromCDATA: true
-        },
-        files: [{
-          expand: true,
-          cwd: '<%= yeoman.dist %>',
-          src: ['*.html'],
-          dest: '<%= yeoman.dist %>'
-        }]
-      }
-    },
-
-    ngtemplates: {
-      dist: {
-        options: {
-          module: 'climbingMemo',
-          htmlmin: '<%= htmlmin.dist.options %>',
-          usemin: 'scripts/scripts.js'
-        },
-        cwd: '<%= yeoman.app %>',
-        src: 'views/**/*.html',
-        dest: '.tmp/templateCache.js'
-      }
-    },
-
-    // ng-annotate tries to make the code safe for minification automatically
-    // by using the Angular long form for dependency injection.
-    ngAnnotate: {
-      dist: {
-        files: [{
-          expand: true,
-          cwd: '.tmp/concat/scripts',
-          src: '*.js',
-          dest: '.tmp/concat/scripts'
-        }]
-      }
-    },
-
-    // Replace Google CDN references
-    cdnify: {
-      dist: {
-        html: ['<%= yeoman.dist %>/*.html']
-      }
-    },
-
-    // Copies remaining files to places other tasks can use
-    copy: {
-      dist: {
-        files: [{
-          expand: true,
-          dot: true,
-          cwd: '<%= yeoman.app %>',
-          dest: '<%= yeoman.dist %>',
-          src: [
-            '*.{ico,png,txt}',
-            '**/*.js',
-            '**/*.html',
-            'images/**/*.{webp}',
-            'styles/fonts/**/*.*'
-          ]
-        }, {
-          expand: true,
-          cwd: '.tmp/images',
-          dest: '<%= yeoman.dist %>/images',
-          src: ['generated/*']
-        }, {
-          expand: true,
-          cwd: 'bower_components/bootstrap/dist',
-          src: 'fonts/*',
-          dest: '<%= yeoman.dist %>'
-        }, {
-          expand: true,
-          cwd: 'bower_components/components-font-awesome',
-          src: 'fonts/*',
-          dest: '<%= yeoman.dist %>'
-        }]
-      },
-      styles: {
-        expand: true,
-        cwd: '<%= yeoman.app %>/styles',
-        dest: '.tmp/styles/',
-        src: '**/*.css'
-      }
-    },
-
-    manifest: {
-      generate: {
-        options: {
-          basePath: '<%= yeoman.dist %>',
-          cache: [
-            'index.html',
-            'fonts/glyphicons-halflings-regular.woff2',
-            'fonts/glyphicons-halflings-regular.woff',
-            'fonts/glyphicons-halflings-regular.ttf',
-            'fonts/fontawesome-webfont.ttf?v=4.4.0',
-            'fonts/fontawesome-webfont.woff?v=4.4.0',
-            'fonts/fontawesome-webfont.woff2?v=4.4.0',
-            'styles/main.css',
-            'styles/vendor.css',
-            'scripts/vendor.js ',
-            'scripts/scripts.js'
-          ],
-          network: ['*'],
-          fallback: ['/ /offline.html'],
-          exclude: ['manifest.appcache'],
-          preferOnline: true,
-          verbose: true,
-          timestamp: true,
-          hash: true,
-          master: ['index.html']
-        },
-        src: [
-          'components/**/*.js',
-          'scripts/**/*.js',
-          'styles/**/*.css',
-          'views/**/*.html'
-        ],
-        dest: 'manifest.appcache'
-      }
-    },
-
-    // Run some tasks in parallel to speed up the build process
-    concurrent: {
-      server: [
-        'compass:server'
-      ],
-      test: [
-        'compass'
-      ],
-      dist: [
-        'compass:dist',
-        'imagemin',
-        'svgmin'
-      ]
-    },
-
-    // Test settings
-    karma: {
-      unit: {
-        configFile: 'test/karma.conf.js',
-        singleRun: true
-      }
-    },
-
-    protractor: {
-      options: {
-        keepAlive: false,
-        noColor: false
-      },
-      local: {
-        options: {
-          configFile: 'test/e2e/protractor.conf.js'
-        }
-      },
-      sauceLab: {
-        options: {
-          configFile: 'test/e2e/protractorSauceLab.conf.js',
-          args: {
-            sauceUser: process.env.SAUCE_USERNAME,
-            sauceKey: process.env.SAUCE_ACCESS_KEY
-          }
         }
       }
     },
-
-    coveralls: {
-      options: {
-        coverageDir: 'coverage',
-        force: true,
-        recursive: true
+    jitGrunt: {
+      staticMappings: {
+        useminPrepare: 'grunt-usemin',
+        ngtemplates: 'grunt-angular-templates',
+        cdnify: 'grunt-google-cdn',
+        coveralls: 'grunt-karma-coveralls',
+        protractor: 'grunt-protractor-runner',
+        versioncheck: 'grunt-version-check'
       }
     }
   })
 
+  // config.uglify = {
+  //   dist: {
+  //     files: {
+  //       'dist/scripts/scripts.js': [
+  //         'dist/scripts/scripts.js'
+  //       ]
+  //     }
+  //   }
+  // }
+
+  grunt.initConfig(config)
 
   grunt.registerTask('serve', 'Compile then start a connect web server', function(target) {
     if (target === 'dist') {
@@ -566,11 +65,6 @@ module.exports = function(grunt) {
     ])
   })
 
-  grunt.registerTask('server', 'DEPRECATED TASK. Use the "serve" task instead', function(target) {
-    grunt.log.warn('The `server` task has been deprecated. Use `grunt serve` to start a server.')
-    grunt.task.run(['serve:' + target])
-  })
-
   grunt.registerTask('help', 'Display some helpful information if no task provided', function() {
     this.async()
     grunt.util.spawn({
@@ -578,61 +72,4 @@ module.exports = function(grunt) {
         grunt.log.writelns(r, '\n')
       })
   })
-
-  grunt.registerTask('e2e-test', [
-    'connect:test',
-    'protractor:local'
-  ])
-
-  grunt.registerTask('unit-test', [
-    'jscs',
-    'jshint',
-    'karma',
-    'coveralls'
-  ])
-
-  grunt.registerTask('full-test', [
-    'clean:server',
-    'wiredep',
-    'concurrent:test',
-    'autoprefixer',
-    'unit-test',
-    'connect:test',
-    'protractor:sauceLab'
-  ])
-
-  grunt.registerTask('test', [
-    'clean:server',
-    'wiredep',
-    'concurrent:test',
-    'autoprefixer',
-    'unit-test',
-    'e2e-test'
-  ])
-
-  grunt.registerTask('build', [
-    'clean:dist',
-    'wiredep',
-    'useminPrepare',
-    'concurrent:dist',
-    'autoprefixer',
-    'ngtemplates',
-    'concat',
-    'ngAnnotate',
-    'copy:dist',
-    'cdnify',
-    'cssmin',
-    // 'uglify',
-    // 'filerev',
-    'usemin',
-    'htmlmin',
-    'manifest'
-  ])
-
-  grunt.registerTask('default', [
-    'newer:jshint',
-    'newer:jscs',
-    'test',
-    'build'
-  ])
 }
