+++ grunt/cdnify.js
@@ -1,0 +1,5 @@
+module.exports = {
+  dist: {
+    html: ['dist/*.html']
+  }
+}
