+++ grunt/clean.js
@@ -1,0 +1,13 @@
+module.exports = {
+  dist: {
+    files: [{
+      dot: true,
+      src: [
+        '.tmp',
+        '<%= path.dist %>/**/*',
+        '!<%= path.dist %>/.git**/*'
+      ]
+    }]
+  },
+  server: '.tmp'
+}
