+++ grunt/concurrent.js
@@ -1,0 +1,13 @@
+module.exports = {
+  server: [
+    'compass:server'
+  ],
+  test: [
+    'compass'
+  ],
+  dist: [
+    'compass:dist',
+    'imagemin',
+    'svgmin'
+  ]
+}
