+++ grunt/coveralls.js
@@ -1,0 +1,7 @@
+module.exports = {
+  options: {
+    coverageDir: 'coverage',
+    force: true,
+    recursive: true
+  }
+}
