+++ grunt/cssmin.js
@@ -1,0 +1,9 @@
+module.exports = {
+  dist: {
+    files: {
+      '<%= path.dist %>/styles/main.css': [
+        '.tmp/styles/**/*.css'
+      ]
+    }
+  }
+}
