+++ grunt/htmlmin.js
@@ -1,0 +1,16 @@
+module.exports = {
+  dist: {
+    options: {
+      collapseWhitespace: true,
+      conservativeCollapse: true,
+      collapseBooleanAttributes: true,
+      removeCommentsFromCDATA: true
+    },
+    files: [{
+      expand: true,
+      cwd: '<%= path.dist %>',
+      src: ['*.html'],
+      dest: '<%= path.dist %>'
+    }]
+  }
+}
