+++ grunt/jscs.js
@@ -1,0 +1,11 @@
+module.exports = {
+  src: [
+    'Gruntfile.js',
+    '<%= path.app %>/components/**/*.js',
+    '<%= path.app %>/scripts/**/*.js',
+    '<%= path.test %>/**/*.js'
+  ],
+  options: {
+    config: '.jscsrc'
+  }
+}
