+++ grunt/karma.js
@@ -1,0 +1,6 @@
+module.exports = {
+  unit: {
+    configFile: '<%= path.karmaConf %>',
+    singleRun: true
+  }
+}
