+++ grunt/ngAnnotate.js
@@ -1,0 +1,10 @@
+module.exports = {
+  dist: {
+    files: [{
+      expand: true,
+      cwd: '.tmp/concat/scripts',
+      src: '*.js',
+      dest: '.tmp/concat/scripts'
+    }]
+  }
+}
