+++ grunt/ngtemplates.js
@@ -1,0 +1,12 @@
+module.exports = {
+  dist: {
+    options: {
+      module: 'climbingMemo',
+      htmlmin: '<%= options.htmlmin %>',
+      usemin: 'scripts/scripts.js'
+    },
+    cwd: 'app/',
+    src: 'views/**/*.html',
+    dest: '.tmp/templateCache.js'
+  }
+}
