+++ grunt/svgmin.js
@@ -1,0 +1,10 @@
+module.exports = {
+  dist: {
+    files: [{
+      expand: true,
+      cwd: '<%= path.app %>/images',
+      src: '**/*.svg',
+      dest: '<%= path.dist %>/images'
+    }]
+  }
+}
