+++ grunt/useminPrepare.js
@@ -1,0 +1,15 @@
+module.exports = {
+  html: '<%= path.app %>/index.html',
+  options: {
+    dest: 'dist/',
+    flow: {
+      html: {
+        steps: {
+          js: ['concat'],
+          css: ['cssmin']
+        },
+        post: {}
+      }
+    }
+  }
+}
