+++ grunt/versioncheck.js
@@ -1,0 +1,5 @@
+module.exports = {
+  options: {
+    hideUpToDate: false
+  }
+}
