+++ package.json
@@ -34,14 +34,15 @@
     "grunt-version-check": "^0.2.1",
     "grunt-wiredep": "^2.0.0",
     "gzippo": "^0.2.0",
-    "jit-grunt": "^0.9.1",
     "jscoverage": "~0.5.9",
     "jshint-stylish": "^2.0.1",
     "karma-chrome-launcher": "^0.2.0",
     "karma-coverage": "^0.4.2",
     "karma-jasmine": "^0.3.6",
+    "grunt-jsdoc": "^1.1.0",
     "karma-ng-html2js-preprocessor": "^0.1.2",
     "karma-phantomjs-launcher": "*",
+    "load-grunt-config": "^0.19.1",
     "morgan": "^1.6.1",
     "time-grunt": "^1.0.0"
   },
@@ -55,7 +56,6 @@
   },
   "devDependencies": {
     "codacy-coverage": "^1.1.3",
-    "grunt-jsdoc": "^1.1.0",
     "grunt-protractor-runner": "^3.0.0",
     "ink-docstrap": "^1.0.2"
   }
